+++ custom_components/wnhf/executions/generic_real.py
@@ -92,7 +92,7 @@ class GenericExecutionEngine:
     """
 
     API_VERSION = "1.0"
-    VERSION = "2.1-rc9"
+    VERSION = "2.2-rc10"
     ENABLED_ACTIONS = REAL_EXECUTION_ENABLED_ACTIONS
 
     def __init__(